fix: refactoring by provide/inject
