+++ plugins/provide.ts
@@ -1,6 +1,8 @@
+import { Context } from '@nuxt/types'
 import { provide } from '@vue/composition-api'
 import { DefaultApolloClient } from '@vue/apollo-composable'
 import ApolloClient from 'apollo-boost'
+import { AuthenticatorKey, Authenticator } from '@/types/firebase/authenticator'
 
 const apolloClient = new ApolloClient({
   uri: 'http://localhost:3000/query',
@@ -13,8 +15,10 @@ const apolloClient = new ApolloClient({
     })
   },
 })
-export default function ({ app }) {
+
+export default function ({ app }: Context) {
   app.setup = () => {
+    provide(AuthenticatorKey, new Authenticator())
     provide(DefaultApolloClient, apolloClient)
   }
 }
