+++ types/firebase/authenticator.ts
@@ -1,5 +1,6 @@
 import firebase from "firebase/app"
 import 'firebase/auth'
+import { InjectionKey } from '@vue/composition-api'
 
 export interface IAuthenticator {
   isSignIn: () => boolean
@@ -7,10 +8,24 @@ export interface IAuthenticator {
   signOut: () => void
 }
 
+export const AuthenticatorKey: InjectionKey<IAuthenticator> = Symbol('IAuthenticator')
+
+const firebaseConfig = {
+  apiKey: 'AIzaSyDoxQU1-q6IrXoo9lPHH8wLBGm4dr5BA2w',
+  authDomain: 'ness-85856.firebaseapp.com',
+  databaseURL: 'https://ness-85856.firebaseio.com',
+  projectId: 'ness-85856',
+  storageBucket: 'ness-85856.appspot.com',
+  messagingSenderId: '410365716960',
+  appId: '1:410365716960:web:01fb9a4a9869527b2e2385',
+  measurementId: 'G-E34BNZV0JP'
+};
+
 export class Authenticator implements IAuthenticator{
   private firebaseAuth: firebase.auth.Auth
-  constructor(auth: firebase.auth.Auth) {
-    this.firebaseAuth = auth
+  constructor() {
+    firebase.initializeApp(firebaseConfig)
+    this.firebaseAuth = firebase.auth()
   }
 
   isSignIn ():boolean {
